Solve conflicts
