+++ .idea/misc.xml
@@ -16,7 +16,9 @@
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.38504180908203134" />
         <entry key="app/src/main/res/layout/activity_sign_up.xml" value="0.38134460449218754" />
         <entry key="app/src/main/res/layout/fragment_map.xml" value="0.32083333333333336" />
+        <entry key="app/src/main/res/layout/fragment_table_view.xml" value="0.45185185185185184" />
         <entry key="app/src/main/res/layout/new_report_item.xml" value="0.35688956433637287" />
+        <entry key="app/src/main/res/menu/main_menu.xml" value="0.33055555555555555" />
         <entry key="app/src/main/res/menu/menu_button_navigation.xml" value="0.5625" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.182" />
         <entry key="app/src/main/res/xml/fileprovider.xml" value="0.5635416666666667" />
