+++ app/src/main/java/com/elmeradrianv/shesafe/fragments/MapFragment.java
@@ -5,20 +5,18 @@ import static com.google.android.gms.location.LocationServices.getFusedLocationP
 import android.Manifest;
 import android.location.Location;
 import android.os.Bundle;
-import android.os.SystemClock;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.BounceInterpolator;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import com.elmeradrianv.shesafe.R;
+import com.elmeradrianv.shesafe.auxiliar.PinAnimation;
 import com.elmeradrianv.shesafe.database.Report;
-import com.elmeradrianv.shesafe.database.TypeOfCrime;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.maps.CameraUpdate;
@@ -26,8 +24,6 @@ import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.BitmapDescriptor;
-import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
@@ -160,64 +156,11 @@ public class MapFragment extends Fragment implements OnMapReadyCallback {
 
     private void showMarker(String title, double latitude, double longitude, int levelOfRisk) {
         LatLng reportLatLng = new LatLng(latitude, longitude);
-        Marker marker;
+
         MarkerOptions markerOptions = new MarkerOptions()
                 .position(reportLatLng)
-                .title(title).icon(getNewIcon(levelOfRisk));
-        marker = map.addMarker(markerOptions);
-        dropPinEffect(marker);
-    }
-
-    private void dropPinEffect(Marker marker) {
-        // Handler allows us to repeat a code block after a specified delay
-        final android.os.Handler handler = new android.os.Handler();
-        final long start = SystemClock.uptimeMillis();
-        final long duration = 1500;
-        // Use the bounce interpolator
-        final android.view.animation.Interpolator interpolator =
-                new BounceInterpolator();
-        // Animate marker with a bounce updating its position every 15ms
-        handler.post(new Runnable() {
-            @Override
-            public void run() {
-                long elapsed = SystemClock.uptimeMillis() - start;
-                // Calculate t for bounce based on elapsed time
-                float t = Math.max(
-                        1 - interpolator.getInterpolation((float) elapsed
-                                / duration), 0);
-                // Set the anchor
-                marker.setAnchor(0.5f, 1.0f + 14 * t);
-
-                if (t > 0.0) {
-                    // Post this event again 15ms from now.
-                    handler.postDelayed(this, 15);
-                } else { // done elapsing, show window
-                    marker.showInfoWindow();
-                }
-            }
-        });
-    }
-
-    private BitmapDescriptor getNewIcon(int levelOfRisk) {
-        BitmapDescriptor icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_yellow);
-        ;
-        switch (levelOfRisk) {
-            case TypeOfCrime.LOW_RISK:
-                icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_yellow);
-                break;
-            case TypeOfCrime.MEDIUM_LOW_RISK:
-                icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_orange);
-                break;
-            case TypeOfCrime.MEDIUM_RISK:
-                icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_red);
-                break;
-            case TypeOfCrime.MEDIUM_HIGH_RISK:
-                icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_blue);
-                break;
-            case TypeOfCrime.HIGH_RISK:
-                icon = BitmapDescriptorFactory.fromResource(R.drawable.ic_ss_marker_purple);
-                break;
-        }
-        return icon;
+                .title(title).icon(PinAnimation.getNewIconWithLevelOfRisk(levelOfRisk));
+        Marker marker = map.addMarker(markerOptions);
+        PinAnimation.dropPinEffect(marker);
     }
 }
